FIXUP: CHROMIUM: sysrq: Added Chrome OS specific x key
Kcr-depends-on-sha: 8e1e49550dc85694abd04d86a8ee36bc98bd8b9e
Kcr-patch: fixups/FIXUP: CHROMIUM: sysrq: Added Chrome OS specific x key.patch

+++ drivers/tty/sysrq.c
@@ -477,7 +477,7 @@ static void sysrq_x_cros_signal_process(char *comm, char *parent, int sig)
 /* how many seconds do we wait for subsequent keypresses after the first */
 #define CROS_SYSRQ_WAIT 20
 
-static void sysrq_handle_cros_xkey(int key)
+static void sysrq_handle_cros_xkey(u8 key)
 {
 	static unsigned long first_jiffies = INITIAL_JIFFIES - CROS_SYSRQ_WAIT * HZ;
 	static unsigned int xkey_iteration;
